I forgot to add the change I mentioned earlier to how type skeletons are represented to the bytecode description as well as the linkcode description, but this should be fixed now.  Sorry if this causes any issues.
1. The module marker is intentionally left out.

2. This is correct - strings in bytecode should not contain a '\0'.  I 
have now updated bytecode.txt and linkcode.txt to mention this, and have 
made the relation between the affected fields more direct.

3. The size of bytecode field should be the length of a machine word.

4. Neededness info is now removed from bytecode and linkcode, and the 
loader sets it to 0.

5. Hidden constants no longer have type environment sizes in bytecode 
and linkcode.

6. The next clause table size is 2 bytes.  The description has been 
fixed.

7. The next clause table should be the same as "predicate definitions 
possibly extending previous ones".

8. While it is correct that the "next clause table", "exportdef 
predicates table", and "local predicates table" are all global, global, 
and local respectively, these marks should be output along with the 
indexes in order to maintain consistency first between the 
representation of constant indexes in the bytecode, and second between 
the representation of the next clause table in the bytecode and 
linkcode, as there may be local constants listed in a next clause table 
in linkcode.

9.  Yes, the 1-byte flag refers to global/local.  I have revised the way 
constant and kind indexes are referred to in the file to be more 
consistent.

10.  Using 1 as a placeholder for the add_imports and remove_imports 
argument should be fine.


git-svn-id: https://teyjus.googlecode.com/svn/new-teyjus@157 da2563f4-8c35-0410-a6f7-ef3e7fa91fdf