Fixed a bug with parsing of list expressions of the form [e1,...,en|e0].
Trimmed unnecessary arguments from a few functions; e.g. parseTerm does 
not use its stack argument and must not return a non-empty stack either.

Added comments in a few places.

a list of terms was not being reversed as needed at some places in the 
grammar description. Fixed this as well.


git-svn-id: https://teyjus.googlecode.com/svn/new-teyjus@299 da2563f4-8c35-0410-a6f7-ef3e7fa91fdf