The "bug fix" in parseTerm was actually unnecessary because the case would never arise. Reverted back to the old code that was simpler but added a comment explaining why it is okay.
Added documentation explaining the arguments of parseTerm. One place here is
left for Zach to fill---I don't know what parsingtoplevel is about.

Comments of this kind need to be added for all the functions that have several
arguments that are not necessarily self-explanatory. Perhaps Zach can help/do
this since he is most familiar with the code? I will also take a try as time 
permits.


git-svn-id: https://teyjus.googlecode.com/svn/new-teyjus@300 da2563f4-8c35-0410-a6f7-ef3e7fa91fdf