Fixed a bug in the function getNewVarsItType in the file types.ml. This function was assuming for some reason that it would never see the ErrorType. However, when I look at the code for translateTerm, it has a call to Types.getNewVarsInTypeMol *even when there may have been errors* in the input term. specifically even when there may have been type errors.
I have fixed only one error of this kind. I see that the general structure
of translateTerm is such that it follows the same path whether or not errors
have been encountered in the process of parsing and analyzing a term. This
seems reasonable but then all the functions used in the code must have the 
ability to deal with error terms and types. Zach needs to look carefully
at the code in the parse.ml and types.ml files and make sure that this kind 
of problem situation does not exist for *any* of the functions defined in 
them.


git-svn-id: https://teyjus.googlecode.com/svn/trunk@559 da2563f4-8c35-0410-a6f7-ef3e7fa91fdf