Fixed term printing in Absyn.string_of_term: the error was that application arguments were being printed in the wrong term context.
Changed the interface to Parse.translateTerm and Parse.translateClause so that they return options; None is returned if an error is encountered.  Additionally, changed the treatment of Errormsg.anyErrors in the context of parsing so that it is false on entry and exit to the two aforementioned functions.  Then, changed the two places that referenced these functions so that they fit with the new interface.

Changed the Visual Studio solution so that the linker is built.  This required renaming a file (/source/linker/message.*) so that compilation could proceed correctly due to a Visual Studio idiosyncrasy.

Changed the README about compiling Teyjus on Windows to reflect the changes to the solution.


git-svn-id: https://teyjus.googlecode.com/svn/trunk@578 da2563f4-8c35-0410-a6f7-ef3e7fa91fdf