Fixed a bug related to accumulated modules using accum_sig in their signatures for kinds and constants.
Cleaned up printing of skeleton type variables.

Added some error handling so that exceptions when opening files are handled more cleanly in disassembler.

Added code to the simulator and linker to eliminate spurious warnings related to falling off the end of a function.  Added code to the table and instruction generators to get rid of some warnings as well.

Changed the VS solution so that it handles cleaning up after ocamlopt, and changed the README to reflect this.

Updated the test outputs that changed due to the change in printing types.


git-svn-id: https://teyjus.googlecode.com/svn/trunk@651 da2563f4-8c35-0410-a6f7-ef3e7fa91fdf