Created a new toplevel tool called tjdepend. This tool takes a list of module files and displays the dependencies required to build the *.lp and *.lpo files in a format suitable for inclusion in Makefiles. I have updated examples/pcf/Makefile to use tjdepend.
Zach, please update the visual studio builds to include this new tool. The frontend is in source/front/dependfront.ml and it depends on a few files from the compiler directory. The exact dependencies can be found in source/OMakefile.

Gopalan, please look at source/front/dependfront.ml and ensure that I have correctly implemented the dependency algorithm described in your email. Is it true that use_sig dependencies are ignored?

On a side note, I put all the logic for tjdepend into source/front/dependfront.ml since there was not enough to justify an entirely new directory.


git-svn-id: https://teyjus.googlecode.com/svn/trunk@653 da2563f4-8c35-0410-a6f7-ef3e7fa91fdf