Created issue directories for bugs in the renaming branch
I have created the directories issue1 and issue2 which contain examples of bugs in the renaming branch. If you run 'tjcc root' in an issue directory (or more generally run 'make') you should see the erroneous behavior. I have also documented the errors in the sig/mod files.

Putting bugs into these types of toplevel directories may not be the best practice, but it should be okay for exchanging this type of information, particularly on a branch. We should find a way to migrate such things into automated test cases.

git-svn-id: https://teyjus.googlecode.com/svn/branches/renaming@972 da2563f4-8c35-0410-a6f7-ef3e7fa91fdf