Updated renaming function, detects range colission
Minor changes to the renaming function logic, now 
uses a reverse hash table in order to ensure the 
renaming function is bijective before passing it
on to the rest of the program. Removed print
statements because they were bugging me (Sorry 
Peter :p). The change in bytecode.ml is a simple
update for forward compatibility: String.set is 
functionally equivalent to Bytes.set, and was
just a deprecated alias.

Strangely, an unrelated linker test is now failing.
Will look into this shortly.


git-svn-id: https://teyjus.googlecode.com/svn/branches/RenamingRedux@1158 da2563f4-8c35-0410-a6f7-ef3e7fa91fdf