Fix for type skeletons not being written to bytecode
Changed logic in spitcode.ml to properly write all type
skeletons passed into it by the Codegen module. Previously,
there was a comparison to see if a given type skeleton was a
duplicate of one already written: however, all duplicates
had already been removed, and this check was causing
problems when there were copies of the same type skeleton
already in the list.

Minor indentation changes to codegen.ml and absyn.ml


git-svn-id: https://teyjus.googlecode.com/svn/branches/RenamingRedux@1169 da2563f4-8c35-0410-a6f7-ef3e7fa91fdf