Simplified renaming logic, sanitized renaming, added tests
Internally, include statements (i.e. {kind a}) are
represented the same way as {kind a => a}. This greatly
simplified the code handling renaming, and now reflects our
specification that statements akin to those two should be
considered equivalent. 

Also, before a signature's accumulated signatures are
compiled/translated, they're put through a sanitation
function which:
 1) Prints out warnings if there are multiple equivalent
    renaming statements, i.e. {kind a, kind a => a}.
 2) Removes the signature from the accumulated signature
    list if the renaming directive is empty ({}), and
    prints a warning.

Currently only does so for signature accumulation - a
generalization which applies to module renaming is
forthcoming in the next commit.



git-svn-id: https://teyjus.googlecode.com/svn/branches/RenamingRedux@1178 da2563f4-8c35-0410-a6f7-ef3e7fa91fdf