Implemented teyjus-inferior-mode
Three kinds of functionality are implemented

1. `teyjus-run` starts tjsim as a comint inferior process

2. `teyjus-path-add-current-buffer` and
`teyjus-path-remove-current-buffer` respectively add and remove the
location of the file being visited to TJPATH.

3. `teyjus-load-buffer` is composed of 1, 2 along with the existing
`teyjus-compile`. It adds the location of the current module to
TJPATH, saves the module files, and compiles, links, and loads them
into an inferior process. It will start a new process or restart an
existing process as needed.