Change how we check if a file is writable
Since dune created artifacts are created without write permission, a
generated file may try to rewrite files that already exist and be denied
permission to do so. Instead of checking for permission denied
explicitly, check to see if the file already exists.

In either case, when a generated file is changed, the entire build
should be rebuilt to ensure the newest versions are present throughout
the build.